# services/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sling Scheduler API
  version: 1.0.0
servers:
  - url: https://<cloud-run-url>
paths:
  /healthz:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /api/schedule:
    get:
      summary: Get schedule for a single date
      parameters:
        - in: query
          name: date
          required: true
          schema:
            type: string
            pattern: '^\\d{4}-\\d{2}-\\d{2}$'
      responses:
        '200':
          description: Schedule payload
        '400':
          description: Invalid date
  /api/shifts/{occurrenceId}:
    put:
      summary: Update one occurrence only
      parameters:
        - in: path
          name: occurrenceId
          required: true
          schema:
            type: string
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                startTime:
                  type: string
                  example: '11:30'
                endTime:
                  type: string
                  example: '16:30'
                status:
                  type: string
                  enum: [published, planning]
      responses:
        '200':
          description: Updated
        '400':
          description: Validation error
        '409':
          description: Idempotency key conflict or in-progress reservation
  /api/shifts/bulk:
    post:
      summary: Bulk update occurrences with partial success support
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                updates:
                  type: array
                  items:
                    type: object
                    properties:
                      occurrenceId:
                        type: string
                      startTime:
                        type: string
                      endTime:
                        type: string
                      status:
                        type: string
                        enum: [published, planning]
      responses:
        '200':
          description: ok or partial_success
        '409':
          description: full failure or idempotency key conflict/in-progress reservation